Blueprint for order part of application
