+++ app/order/__init__.py
@@ -1,0 +1,5 @@
+from flask import Blueprint
+
+bp = Blueprint('order', __name__)
+
+from app.order import routes, forms
