+++ app/templates/base.html
@@ -26,7 +26,7 @@
                 <ul class="nav navbar-nav">
                     <li><a href="{{ url_for('main.index') }}">Home</a></li>
                     <li><a href="{{ url_for('customer.customers') }}">Customers</a></li>
-                    <li><a href="{{ url_for('list_orders') }}">Orders</a></li>
+                    <li><a href="{{ url_for('order.list_orders') }}">Orders</a></li>
                     <li><a href="{{ url_for('list_controllers') }}">Controllers</a></li>
                 </ul>
                 <ul class="nav navbar-nav navbar-right">
