+++ app/templates/main/index.html
@@ -15,7 +15,7 @@
     <div class="row text-center">
         <div class="col-lg-offset-4">
             <div class="col-lg-3">
-                <a class="btn btn-info btn-block" href="{{ url_for('list_orders') }}">List Orders</a>
+                <a class="btn btn-info btn-block" href="{{ url_for('order.list_orders') }}">List Orders</a>
             </div>
         </div>
         <div class="col-lg-2">
