+++ app/templates/order/orders.html
@@ -3,7 +3,7 @@
 
 {% block app_content %}
     <h1>{{ title }}:</h1>
-    <a class="btn btn-primary" href="{{ url_for('create_order', customer_id=customer_id) }}">New Order</a>
+    <a class="btn btn-primary" href="{{ url_for('order.create_order', customer_id=customer_id) }}">New Order</a>
     <table class="table table-striped">
         <thead>
         <tr>
@@ -15,7 +15,7 @@
         </thead>
         {% for order in orders %}
             <tr>
-                <td><a href="{{ url_for('detail_order', number=order.id) }}">{{ order.id }} Details</a></td>
+                <td><a href="{{ url_for('order.detail_order', number=order.id) }}">{{ order.id }} Details</a></td>
                 <td>{{ order.ride.capitalize() }}</td>
                 {% if order.two_forty and order.three_twenty %}
                     <td>Both</td>
